cleaned code of SearchRecipeViewModel, added ListRecipeViewModel
