+++ app/src/main/java/com/example/cookit/MainActivity.kt
@@ -21,7 +21,9 @@ import androidx.navigation.NavType
 import androidx.navigation.navArgument
 import com.example.cookit.screens.listRecipes.ListRecipesScreen
 import com.example.cookit.screens.components.Recipe
+import com.example.cookit.screens.createRecipe.CreateRecipeViewModel
 import com.example.cookit.screens.editRecipe.EditRecipeScreen
+import com.example.cookit.screens.listRecipes.ListRecipesViewModel
 import com.example.cookit.screens.searchRecipe.FilterIngredientsScreen
 import com.example.cookit.screens.searchRecipe.SearchRecipeScreen
 //import com.example.cookit.screens.searchRecipe.FilterIngredientsScreen
@@ -80,7 +82,10 @@ fun App() {
             FilterIngredientsScreen(navController, recipeViewModel)
         }
 
-        composable("createRecipe") { CreateRecipeScreen(navController) }
+        composable("createRecipe") {
+            val createRecipeViewModel: CreateRecipeViewModel = viewModel()
+            CreateRecipeScreen(navController, createRecipeViewModel)
+        }
 
         composable(
             route = "showRecipe/{recipeJson}",
@@ -107,6 +112,9 @@ fun App() {
                 }
             )
         ) { backStackEntry ->
+
+            val listRecipesViewModel: ListRecipesViewModel = viewModel()
+            
             val recipeIdListJson = backStackEntry.arguments?.getString("recipeIdList")
             val isFavorites = backStackEntry.arguments?.getBoolean("isFavorites") ?: false
             val type = backStackEntry.arguments?.getString("type") ?: ""
@@ -116,6 +124,7 @@ fun App() {
             } ?: emptyList()
 
             ListRecipesScreen(
+                viewModel = listRecipesViewModel,
                 navController = navController,
                 recipeIds = recipeIdList,
                 isFavorites = isFavorites,
